feat: scaffold model viewer refactor
