+++ protected/js/model-viewer/viewer/index.js
@@ -1,0 +1,3 @@
+export function createViewer() {
+  //
+}
