+++ protected/views/site/_model_viewer.php
@@ -65,183 +65,9 @@ $files = array_map(function ($filename) use ($assetsUrl) {
   </div>
 </div>
 
-<script type="module">
-  import * as THREE from 'three';
-  import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-  import { STLLoader } from 'three/addons/loaders/STLLoader.js';
 
-  let camera, scene, renderer, controls;
-
-  const height = 512;
-  const width = 992;
-  const aspectRatio = width / height;
-
-
-  $(document).ready(function () {
-    const loadingOverlay = $('.js-loading-overlay');
-    const playButtonOverlay = $('.js-play-button-overlay');
-    const playButton = $('.js-play-button');
-    const errorDisplay = $('.js-error-display');
-    const modelSelector = $('.js-model-selector');
-
-    const createModelStateProxy = () => {
-      const state = {
-        loading: false,
-        loaded: false,
-        error: null,
-      };
-
-      return new Proxy(state, {
-        set: function (target, property, value) {
-          console.log('set', target, property, value);
-          target[property] = value;
-
-          if (property === 'loading') {
-            if (value) {
-              showLoadingOverlay();
-              hidePlayButtonOverlay();
-            } else {
-              hideLoadingOverlay();
-            }
-          }
-
-          if (property === 'loaded') {
-            if (value) {
-              hideLoadingOverlay();
-              hidePlayButtonOverlay();
-            } else {
-              showPlayButtonOverlay();
-            }
-          }
-
-          if (property === 'error') {
-            if (value) {
-              setError(value);
-            }
-          }
-
-          return true;
-        }
-      });
-    };
-
-    const modelState = createModelStateProxy();
-
-    function showLoadingOverlay() {
-      loadingOverlay.show();
-    }
-
-    function hideLoadingOverlay() {
-      loadingOverlay.hide();
-    }
-
-    function showPlayButtonOverlay() {
-      playButtonOverlay.show();
-    }
-
-    function hidePlayButtonOverlay() {
-      playButtonOverlay.hide();
-    }
-
-    function setError(message) {
-      if (message) {
-        errorDisplay.addClass('alert', 'alert-danger');
-        errorDisplay.find('.error-text').text(message);
-      } else {
-        errorDisplay.removeClass('alert', 'alert-danger');
-      }
-    }
-
-    // handle three.js setup
-    function initThree() {
-      scene = new THREE.Scene();
-      camera = new THREE.PerspectiveCamera(40, aspectRatio, 0.001, 1000);
-      renderer = new THREE.WebGLRenderer({
-        canvas: document.getElementById('3d-model-canvas'),
-        antialias: true
-      });
-      renderer.setSize(width, height);
-      renderer.setPixelRatio(window.devicePixelRatio);
-      renderer.setClearColor(0xf0f0f0);
-
-      controls = new OrbitControls(camera, renderer.domElement);
-      controls.enableDamping = true;
-      controls.dampingFactor = 0.05;
-      controls.screenSpacePanning = true;
-
-      controls.zoomSpeed = 1.2;
-
-      camera.position.set(5, 5, 5);
-      controls.target.set(0, 0, 0);
-      controls.update();
-
-      $(window).on('resize', onWindowResize);
-    }
-
-    function onWindowResize(event) {
-      renderer.setSize(width, height);
-      camera.aspect = width / height;
-      camera.updateProjectionMatrix();
-    }
-
-    async function loadSTLModel(url) {
-      const loader = new STLLoader();
-      const geometry = await loader.loadAsync(url);
-      geometry.center();
-      const material = new THREE.MeshNormalMaterial({
-        flatShading: true
-      });
-      const mesh = new THREE.Mesh(geometry, material);
-
-      // Scale model to fit view
-      const box = new THREE.Box3().setFromObject(mesh);
-      const size = box.getSize(new THREE.Vector3()).length();
-      const scale = 5 / size;
-      mesh.scale.set(scale, scale, scale);
-
-      scene.add(mesh);
-
-      function animate() {
-        requestAnimationFrame(animate);
-        controls.update();
-        renderer.render(scene, camera);
-      }
-      animate();
-    }
-
-    async function loadModel() {
-      modelState.loaded = false;
-      modelState.loading = true;
-      try {
-        initThree();
-        const modelUrl = modelSelector.val();
-        const extension = modelUrl.split('.').pop();
-        if (extension.toLowerCase() === 'stl') {
-          await loadSTLModel(modelUrl);
-        }
-      } catch (error) {
-        console.error(error);
-        modelState.error = error;
-      } finally {
-        modelState.loading = false;
-        modelState.loaded = true;
-      }
-    }
-
-    function initUi() {
-      loadingOverlay.hide();
-      playButtonOverlay.show();
-      errorDisplay.hide();
-      modelSelector.on('change', loadModel);
-      playButton.on('click', loadModel);
-    }
-
-    initUi();
-
-    $(window).on('beforeunload', () => {
-      modelSelector.off('change', loadModel);
-      playButton.off('click', loadModel);
-    })
-  })
-
-</script>
+<?php
+Yii::app()->assetManager->forceCopy = YII_DEBUG;
+$jsDir = Yii::getPathOfAlias('application.js.model-viewer');
+$jsUrl = Yii::app()->assetManager->publish($jsDir);
+Yii::app()->clientScript->registerScriptFile($jsUrl . '/index.js', CClientScript::POS_END, ['type' => 'module']);
